Squashed the multiples bug and revealed another.
The multiples but was caused by a fencepost error when sharding.
Shards were being geenrated 1 to tall and 1 to wide on the right and top respectively.

The new bug involves, likely, the failure to examine all of the cuboids involved in intersection correctly.
Likely this lies somewhere in the node-fitting algorithm.