Altered the overlap tests to not create permenant nodes in the tree.
It does this by not creating a "downlink" from the parent node. Since most of the algorithms
deal with the lowest nodes and work up except search, delete, and insertion the "upward" links
are enough to have the sharding, etc work flawlessly and the nodes that were created get GC'ed at the end.

This stops unlimited tree growth through intersection searches alone and, as always, keeps the tree minimal.